Split each class into its own file
